constant-propagator: more intelligent constant propagation
Important for code that involves loops, etc. Instead of blindly throwing an
error when an unsupported construct is encountered, split the code into two
halves, with the first half containing everything up to the unsupported
construct (e.g., the start of a loop). Then only perform constant propagation
in the first half. Leave everything in the second halve untouched.